stylize.php: Don't add whitespace in strange places
For example, don't change this:

    $a = $b ?: $c->{ $d };

to this:

    $a = $b ? : $c-> { $d } ;

Specifically, do not add a space after '}' or between '?' and ':'
or '->' and '{'. This can be done merely by editing the tables.

Also, add a space on both sides of '?:' by adding a space after ':'.
As before, this is far from perfect because of case labels, and I
would like to keep more extensive code changes separate from these
simple table changes.

Change-Id: I0eb6e3dd9e6eb35ef37478a59cca1dc3f8aadaf3