Add pre-commit hook for automatically fixing code style issues.
This is a git pre-commit hook, desigend to be run by the git commit command.
The purpose of this script is to run phpcs and phpcbf before the commit
is created, to detect and fix code style issues. If not all issues could
be fixed, the commit is aborted.

This script is designed for use on the mediawiki core repo. It may need
adjustments to work on other repos.

To activate this script for a git repo, place it in the .git/hooks directory
and rename it to "pre-commit". A symlink would also work. If you have
multiple scripts you want to run before every commit, create a script called
"pre-commit", and call each script from there.

Change-Id: I35222d96f442ceaf84cd05322f4a912a169ec16d